requirements.txt
# SupplySense core dependencies
pip>=23.0
setuptools>=68.0

# Core
pandas
numpy
statsmodels
prophet
streamlit

# Visualization
matplotlib
plotly

# ML expansion
scikit-learn

# Utility
python-dotenv
joblib
pytest

